Reorganized types to their own folder, reduced total number of files
